+++ R/addZeroCatch.R
@@ -10,7 +10,7 @@
 #' 
 #' One should test the results of this function by creating a frequency table of the \code{eventvar} or \code{specvar}. In either case, the table should contain the same value in each cell of the table. See the examples.
 #'
-#' @note An error will be returned if either \code{specvar} or \code{eventvar} are factors with any \code{NA} levels. This usually arises if the data.frame was subsetted/filtered prior to using \code{addZeroCatch}. See \code{\link{filterD}} or \code{\link[base]{droplevels}} for descriptions of how to drop unused levels.
+#' @note An error will be returned if either \code{specvar} or \code{eventvar} are factors with any \code{NA} levels. This usually arises if the data.frame was subsetted/filtered prior to using \code{addZeroCatch}. See \code{\link[base]{droplevels}} for descriptions of how to drop unused levels.
 #' 
 #' @param df A data.frame that contains the capture summary data as described in the details.
 #' @param eventvar A string for the variable that identifies unique capture events.